+++ src/pylingdocs/data/format_templates/html/basic/hooks/post_gen_project.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 from shutil import copy
 
-
 try:
     from importlib.resources import files  # pragma: no cover
 except ImportError:  # pragma: no cover