+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * multipart LaTeX examples
 
+### Changed
+
+* `builders` settings is now named `build`
+
 ## [0.0.12] - 2023-07-02
 
 ### Added